Thu, Jul 4, 2024, 9:08 PM +05:30

+++ world/rough2.c
@@ -7,11 +7,11 @@
 
 int main(void)
 {
-    float c1 = 72;
+    int c1 = 72;
     int c2 = 73;
     int c3 = 33;
 
     char h = c1/2;
 
-    printf("%i%c%c%i\n", c1, c2, c3,h);
+    printf("%f%c%c%i\n", c1, c2, c3,h);
 }
